+++ src/main/java/gwasuwonshot/tutice/lesson/controller/PaymentRecordController.java
@@ -6,8 +6,8 @@ import gwasuwonshot.tutice.common.exception.SuccessStatus;
 import gwasuwonshot.tutice.common.module.DateAndTimeConvert;
 import gwasuwonshot.tutice.common.resolver.userIdx.UserIdx;
 import gwasuwonshot.tutice.lesson.dto.request.UpdatePaymentRecordRequest;
-import gwasuwonshot.tutice.lesson.dto.response.getPaymentRecord.GetPaymentRecordResponse;
-import gwasuwonshot.tutice.lesson.dto.response.GetPaymentRecordCycleResponse;
+import gwasuwonshot.tutice.lesson.dto.response.getPaymentRecordByLesson.GetPaymentRecordResponse;
+import gwasuwonshot.tutice.lesson.dto.response.getPaymentRecordCycle.GetPaymentRecordCycleResponse;
 import gwasuwonshot.tutice.lesson.service.PaymentRecordService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
