+++ src/main/java/gwasuwonshot/tutice/lesson/dto/response/createLesson/CreateLessonResponse.java
@@ -1,4 +1,4 @@
-package gwasuwonshot.tutice.lesson.dto.response;
+package gwasuwonshot.tutice.lesson.dto.response.createLesson;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
