+++ src/main/java/gwasuwonshot/tutice/lesson/dto/response/getLessonDetail/GetLessonDetailResponse.java
@@ -1,4 +1,4 @@
-package gwasuwonshot.tutice.lesson.dto.response;
+package gwasuwonshot.tutice.lesson.dto.response.getLessonDetail;
 
 import gwasuwonshot.tutice.common.module.DateAndTimeConvert;
 import gwasuwonshot.tutice.lesson.entity.Lesson;
